MLK-12438 Coverity 30689: STRING_NULL
Bug:
CID 30689 (#1 of 1): String not null terminated (STRING_NULL)
10. string_null_argument: Function fread does not terminate string info.

Solution:
Replace the last char of info.platform[] with a null terminator to avoid
searching behind when passing it to strstr().

info.platform[] is extracted from the header of FW file. It's null
terminatored because its size is large enough to contain platform name.

Add this fix to avoid abnormal header content.

Signed-off-by: Hongzhang Yang <hongzhang.yang@nxp.com>